+++ features/support/paths.rb
@@ -201,6 +201,9 @@ module NavigationHelpers
     when /^the My page$/
       '/my/page'
 
+    when /^the My page personalization page$/
+      "/my/page_layout"
+
     when /^the [mM]y account page$/
       '/my/account'
 
@@ -266,7 +269,7 @@ module NavigationHelpers
 
     when /^the page of the planning element "([^\"]+)" of the project called "([^\"]+)"$/
       planning_element_name = $1
-      planning_element = PlanningElement.find_by_subject(planning_element_name)
+      planning_element = WorkPackage.find_by_subject(planning_element_name)
       "/work_packages/#{planning_element.id}"
 
     when /^the (.+) page (?:for|of) the project called "([^\"]+)"$/