+++ lib/redmine/views/my_page/block.rb
@@ -22,16 +22,17 @@ module Redmine
               gem_name = plugin_id.to_s.gsub('openproject_','openproject-') if plugin_id.to_s.starts_with?('openproject_')
               gem_spec = Gem.loaded_specs[gem_name]
               if gem_spec.nil?
-                ActiveSupport::Deprecation.warn "No Gemspec found for plugin: " + plugin_id
-                  + ", expected gem name to match the plugin name but starting with openproject-"
+                error = "No Gemspec found for plugin: " + plugin_id.to_s \
+                + ", expected gem name to match the plugin name but starting with openproject-"
+                ActiveSupport::Deprecation.warn(error)
                 nil
               else
-                gem_spec.full_gem_path + '/**/my/blocks/_*.{rhtml,erb}'
+                gem_spec.full_gem_path + '/app/views/my/blocks/_*.{rhtml,erb}'
               end
             end.compact
           ).inject({}) do |h,file|
             name = File.basename(file).split('.').first.gsub(/^_/, '')
-            h[name] = name.to_sym
+            h[name] = ("label_" + name).to_sym
             h
           end
         end