+++ doc/CHANGELOG.md
@@ -1,12 +1,22 @@
 # Changelog
 
+* `#1898` Separate action for changing wiki parent page (was same as rename before)
+
+## 3.0.0pre15
+
+* `#1301` Ajax call when logged out should open a popup window
+* `#1351` Generalize Modal Creation
+# `#1557` Timeline Report Selection Not Visible
 * `#1755` Migrate helper-tests for issues into specs for work package
 * `#1766` Fixed bug: Viewing diff of Work Package description results in error 500
+* `#1767` Fixed bug: Viewing changesets results in "page not found"
 * `#1789` Move validation to Work Package
 * `#1800` Add settings to change software name and URL and add additional footer content
 * `#1808` Add option to log user for each request
+* `#1875` Added test steps to reuse steps for my page, my project page, and documents, no my page block lookup at class load time
 * `#1876` Timelines do not show work packages when there is no status reporting
 * `#1896` Moved visibility-tests for issues into specs for workpackages
+* `#1912` Merge column project type with column planning element type
 
 ## 3.0.0pre14